Removed unused ApiQueryAllUsers.php
